Develop

+++ src/modules/users/services/DeleteUserService.ts
@@ -6,6 +6,7 @@ import IUsersRepository from '../repositories/IUsersRepository';
 
 interface Request {
   id: string;
+  tenant_id: string;
 }
 
 @injectable()
@@ -17,11 +18,15 @@ class DeleteUserService {
     /* Anything */
   }
 
-  public async execute({ id }: Request): Promise<void> {
+  public async execute({ id, tenant_id }: Request): Promise<void> {
     const user = await this.usersRepository.findById(id);
 
     if (!user) {
       throw new AppError('Usuário não encontrado.', 404);
+    }
+
+    if (user.tenant_id !== tenant_id) {
+      throw new AppError('Este usuário não faz parte da sua loja.', 401);
     }
 
     await this.usersRepository.delete(id);
